remove core/po
